+++ .gitignore
@@ -39,3 +39,7 @@ yarn-error.log*
 /mts-data-sync
 
 pnpm-lock.yaml
+
+# typescript
+*.tsbuildinfo
+next-env.d.ts
