--- components/Header.js
@@ -1,3 +1,0 @@
-export default function Header({ title }) {
-  return <h1 className="title">{title}</h1>
-}
