+++ package.json
@@ -5,15 +5,13 @@
   "scripts": {
     "dev": "next dev",
     "build": "next build",
-    "export": "next export"
+    "export": "next export",
+    "lint": "next lint"
   },
   "dependencies": {
     "@emotion/react": "^11.10.6",
     "@emotion/styled": "^11.10.6",
     "@mapbox/mapbox-gl-geocoder": "^5.0.1",
-    "@mui/icons-material": "^5.11.11",
-    "@mui/material": "^5.11.14",
-    "@mui/x-data-grid": "^6.2.1",
     "@prisma/client": "^5.1.1",
     "cypress": "^10.0.3",
     "dotenv": "^16.0.3",
@@ -26,9 +24,15 @@
     "react-map-gl": "^7.1.3"
   },
   "devDependencies": {
+    "@types/node": "^20",
+    "@types/react": "^18",
+    "@types/react-dom": "^18",
     "autoprefixer": "^10.4.14",
+    "eslint": "^8",
+    "eslint-config-next": "14.0.3",
     "postcss": "^8.4.27",
     "prisma": "^5.1.1",
-    "tailwindcss": "^3.3.3"
+    "tailwindcss": "^3.3.3",
+    "typescript": "5.3.2"
   }
 }
