--- pages/Landlord_Page.js
@@ -1,8 +1,0 @@
-import LandlordsPage from './getLandlords'
-
-export default function Home() {
-  return (
-    <>
-        < LandlordsPage />
-    </>
-  )}
